x86: dts: ce4100: Fix compatible string for PCF8575 chip
The binding definition for the PCF857x GPIO expanders doesn't mention
a "ti,pcf8575" compatible string.  This is apparently because TI is
only a second source - there is no functional difference between
PCF8575 chips manufactured by TI and NXP, and the same board might be
populated with either depending on availability.

As a practical matter, the gpio-pcf857x driver doesn't match this
string, so this device tree currently won't work.

Signed-off-by: Ben Hutchings <ben@decadent.org.uk>