sound/usb/media: use core routine to initialize media_device
The media core has now a routine to initialize media_device
for USB devices. Use it, instead of doing its own logic,
as it warrants that all USB drivers will behave the same. It
also warrants that the device will get the same data, no matter
if it was initialized initially via snd-usb-audio or via
some other driver, like au0828.

Signed-off-by: Mauro Carvalho Chehab <mchehab@osg.samsung.com>