x86: dts: ce4100: Use defined compatible string for PCF8575 chip
The binding definition for the PCF857x GPIO expanders doesn't mention
a "ti,pcf8575" compatible string.  This is apparently because TI is
only a second source - there is no functional difference between
PCF8575 chips manufactured by TI and NXP, and the same board might be
populated with either depending on availability.

This is not a problem in practice because the I2C core uses
of_modalias_node() before matching drivers and this strips the
manufacturer name.

Signed-off-by: Ben Hutchings <ben.hutchings@codethink.co.uk>