clocksource/arm_global_timer.c: Reduce the rating from 300 to 150
The arm_global_timer clock runs on the CPU clock, and does not correct
for cpufreq scaling. This makes the clock not very suitable as a
clocksource, and basically any clock running on an independent
oscilator is preferable. Few clocksources have a rating over 300,
so this clock usually gets selected as clock source.

On the Zynq-7000 for example, the TTC clock on the chip is preferred
to the global timer, because the TTC can at least compensate for
cpufreq scaling. That makes the TTC (which rates itself 200) a much
better clock source than the ARM global timer.

Reduce the rating to 150 to make systems select the ARM global timer
as a last resort instead of a first choice.

Signed-off-by: Mike Looijmans <mike.looijmans@topic.nl>