qeth: Default to allow promiscuous mode
When a qeth device is in bridge role, one of the ports of an adapter can
ask for promiscuous mode and get it enabled.

The default until now was to not allow user space to enable promiscuous mode
without switching sysfs attributes as well though, diverging from usual
network device semantics.

This patch sets the default to allow promiscuous enablement. That way all
existing tools "just work", albeit only one port of an adapter can be in
promiscuous mode at a given time.

Signed-off-by: Alexander Graf <agraf@suse.de>