cxl: Don't fail initialization if PSL timebase can't be synced
Failure to synchronize the PSL timebase currently prevents the
initialization of the cxl card, thus rendering the card useless. This
is too extreme for a feature which is rarely used, if at all.

This patch still tries to synchronize the PSL timebase when the card
is initialized, but doesn't fail if it can't. Instead, it reports a
status via /sys.

Signed-off-by: Frederic Barrat <fbarrat@linux.vnet.ibm.com>